Fix issues

+++ extension/lib/LogHandler.jsm
@@ -29,7 +29,7 @@ const KILOBYTE = 1024;
 const MEGABYTE = 1024 * KILOBYTE;
 const UPLOAD_LIMIT = 1 * MEGABYTE;
 
-const INITIAL_PADDING = 0.95;
+let padding = 0.95;
 
 
 this.LogHandler = {
@@ -45,17 +45,20 @@ this.LogHandler = {
     const timesinceLastUpload = Date.now() - lastUploadDate;
 
     if (timesinceLastUpload > Config.logSubmissionInterval) {
-      const entriesPingSize = Pioneer.utils.getEncryptedPingSize("online-news-log", 1, entries);
+      let payload = { entries };
+      const entriesPingSize = await Pioneer.utils.getEncryptedPingSize(
+        "online-news-log", 1, payload
+      );
 
       if (entriesPingSize < UPLOAD_LIMIT) {
         // If the ping is small enough, just submit it directly
-        await Pioneer.utils.submitEncryptedPing("online-news-log", 1, entries);
+        await Pioneer.utils.submitEncryptedPing("online-news-log", 1, payload);
+        PrefUtils.setLongPref(UPLOAD_DATE_PREF, Date.now());
       } else {
         // Otherwise, break it into batches below the minimum size
         const reduceRatio = UPLOAD_LIMIT / entriesPingSize;
         const originalEntriesLength = entries.length;
-        let padding = INITIAL_PADDING;
-        let batch;
+        let batch = [];
 
         while (entries.length > 0) {
           const batchSize = Math.floor(originalEntriesLength * reduceRatio * padding);
@@ -64,7 +67,10 @@ this.LogHandler = {
           }
 
           batch = entries.splice(0, batchSize);
-          const batchPingSize = Pioneer.utils.getEncryptedPingSize("online-news-log", 1, batch);
+          payload = { entries: batch };
+          const batchPingSize = await Pioneer.utils.getEncryptedPingSize(
+            "online-news-log", 1, payload
+          );
 
           if (batchPingSize >= UPLOAD_LIMIT) {
             // not small enough, put the batch back in the pool,
@@ -74,19 +80,21 @@ this.LogHandler = {
             continue;
           }
 
-          await Pioneer.utils.submitEncryptedPing("online-news-log", 1, batch);
+          await Pioneer.utils.submitEncryptedPing("online-news-log", 1, payload);
         }
 
         PrefUtils.setLongPref(UPLOAD_DATE_PREF, Date.now());
 
         // Delete the keys that were uploaded
         const lastEntry = batch.pop();
-        const keys = await LogStorage.getAllKeys();
-        for (const key of keys) {
-          if (key <= lastEntry.timestamp) {
-            LogStorage.delete(key);
-          } else {
-            break;
+        if (lastEntry) {
+          const keys = await LogStorage.getAllKeys();
+          for (const key of keys) {
+            if (key <= lastEntry.timestamp) {
+              LogStorage.delete(key);
+            } else {
+              break;
+            }
           }
         }
       }
